requirements.txt
absl-py==2.3.1
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
asyncpg==0.30.0
backoff==2.2.1
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.2
catboost==1.2.8
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.2
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.10
dagshub-annotation-converter==0.1.10
databricks-cli==0.18.0
databricks-sdk==0.58.0
dataclasses-json==0.6.7
debugpy @ file:///C:/b/abs_bf9oo2vhxp/croot/debugpy_1736269476451/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
docker==6.1.3
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastapi==0.116.1
Flask==2.3.3
flatbuffers==25.2.10
fonttools==4.58.5
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-pasta==0.2.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==5.2.0
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1751470649/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///C:/b/abs_7c6fipznw3/croot/jupyter_core_1751991383524/work
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==4.6.0
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
ml_dtypes==0.5.1
mlflow==1.30.0
mlflow-skinny==3.1.1
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.46.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numpy==1.26.4
oauthlib==3.3.1
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt_einsum==3.4.0
optree==0.16.0
packaging==21.3
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathvalidate==3.3.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
plotly==6.2.0
prometheus_client==0.22.1
prometheus_flask_exporter==0.23.2
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.2
protobuf==4.25.8
psutil @ file:///C:/b/abs_b5gv3mn55h/croot/psutil_1736371546320/work
psycopg2==2.9.10
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2022.7.1
pywin32==308
PyYAML==6.0.2
pyzmq @ file:///C:/b/abs_f3yte6j5yn/croot/pyzmq_1734711069724/work
querystring-parser==1.2.4
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
scikeras==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
setuptools==78.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.47.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tornado @ file:///C:/b/abs_7dzpc171lf/croot/tornado_1748956950306/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
waitress==2.1.2
watchfiles==1.1.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.2
xgboost==3.0.2
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
